evaluation/requirements.txt
ragas>=0.2.0
langchain-core>=0.2.0
langchain-openai>=0.1.0
langchain-community>=0.2.0
langchain-qdrant>=0.1.0
qdrant-client>=1.9.0
pymupdf>=1.24.0
langchain-text-splitters>=0.2.0
flashrank>=0.2.0
python-dotenv>=1.0.0
datasets>=2.0.0
pandas>=2.0.0
